feat(sls-rspack): enable invoke local
This will only work for non esm functions when using sls v3.
