+++ libs/serverless-rspack/src/lib/hooks/package/before-create-deployment-artifacts.ts
@@ -1,0 +1,10 @@
+import type { RspackServerlessPlugin } from '../../serverless-rspack.js';
+
+export async function BeforePackageCreateDeploymentArtifacts(
+  this: RspackServerlessPlugin
+) {
+  this.log.verbose('before:package:createDeploymentArtifacts');
+  this.timings.set('before:package:createDeploymentArtifacts', Date.now());
+  await this.bundle(this.functionEntries);
+  await this.pack();
+}
